Split screens into multiple files
Things were a bit messy.

There's `Content` composable that just selects if it should present the login screen or the "main content" screen (which is just a stub right now).
