+++ app/src/main/java/io/github/tiago_vargas/trabalhodeaps/LoginScreen.kt
@@ -1,0 +1,42 @@
+package io.github.tiago_vargas.trabalhodeaps
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import io.github.tiago_vargas.trabalhodeaps.ui.theme.TrabalhoDeApsTheme
+
+@Composable
+fun LoginScreen(onLoginClicked: () -> Unit, modifier: Modifier = Modifier) {
+	Column(
+		modifier = modifier,
+		verticalArrangement = Arrangement.Center,
+		horizontalAlignment = Alignment.CenterHorizontally,
+	) {
+		Button(onClick = onLoginClicked) {
+			Text("Login")
+		}
+	}
+}
+
+@Preview(showBackground = true, showSystemUi = true)
+@Composable
+fun LoginScreenPreview() {
+	TrabalhoDeApsTheme {
+		Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+			LoginScreen(
+				onLoginClicked = {},
+				modifier = Modifier
+					.fillMaxSize()
+					.padding(innerPadding),
+			)
+		}
+	}
+}
